// server/src/core/chains/types/stargaze.testnet.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  Decimal: { input: any; output: any; }
  JSON: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
  NaiveDateTime: { input: any; output: any; }
  UUID: { input: any; output: any; }
};
// ...
export type Event = {
  __typename?: 'Event';
  /** Event action. Some event don't have any */
  action?: Maybe<Scalars['String']['output']>;
  /**
   * Bid Price
   * @deprecated Use the metadata denoms field instead
   */
  bidPrice?: Maybe<Price>;
  /** Attribute bidder if any */
  bidder?: Maybe<Address>;
  blockEventIdx: Scalars['Int']['output'];
  blockHeight: Scalars['Int']['output'];
  blockMessageIdx: Scalars['Int']['output'];
  blockTxIdx: Scalars['Int']['output'];
  /** Attribute buyer if any */
  buyer?: Maybe<Address>;
  contract?: Maybe<Contract>;
  /** Contract Address */
  contractAddr?: Maybe<Scalars['String']['output']>;
  /** Contract Code ID */
  contractCodeId?: Maybe<Scalars['String']['output']>;
  /**
   * Contract Attributes
   * @deprecated Field contract_info is deprecated, use contract { contract_info } instead
   */
  contractInfo?: Maybe<Scalars['JSONObject']['output']>;
  /** Contract Type */
  contractType?: Maybe<Scalars['String']['output']>;
  /** Set the the block time, as UTC without timezone */
  createdAt: Scalars['DateTime']['output'];
  data: Scalars['JSON']['output'];
  /** Event name */
  eventName: Scalars['String']['output'];
  /**
   * Is this event expired. An event can be expired but still valid on chain, for example a bid
   * might be expired but still valid on chain and STARS still be locked. User will have to
   * remove that bid to unlock those STARS, this event will then become invalid (is_valid set to
   * false)
   */
  expired: Scalars['Boolean']['output'];
  /** Internal UUID, might change between when the indexer is reran. */
  id: Scalars['ID']['output'];
  /** False if this event has been superseded by another event */
  isValid: Scalars['Boolean']['output'];
  location: Scalars['Int']['output'];
  message?: Maybe<Message>;
  metadata: Scalars['JSON']['output'];
  /** Attribute operator if any */
  operator?: Maybe<Address>;
  /** Attribute owner if any */
  owner?: Maybe<Address>;
  /**
   * Price
   * @deprecated Use the data denoms field instead
   */
  price?: Maybe<Price>;
  /** Attribute recipient if any */
  recipient?: Maybe<Address>;
  /** Attribute seller if any */
  seller?: Maybe<Address>;
  /** Attribute sender if any */
  sender?: Maybe<Address>;
  transaction?: Maybe<Transaction>;
  /** Transaction Hash */
  txHash?: Maybe<Scalars['String']['output']>;
};
// ...
export type Message = {
  __typename?: 'Message';
  blockHeight: Scalars['Int']['output'];
  /**
   * Transaction Idx
   * Message Idx within the transaction
   */
  blockTxIdx: Scalars['Int']['output'];
  contract?: Maybe<Contract>;
  /** Cosmwasm Method */
  cosmwasmMethod?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['DateTime']['output'];
  /** Message Attributes */
  data: Scalars['JSON']['output'];
  /** Careful: this is only returning the last 100 events */
  events?: Maybe<Array<Event>>;
  /** Internal UUID, might change between when the indexer is reran. */
  id: Scalars['ID']['output'];
  idx: Scalars['Int']['output'];
  transaction?: Maybe<Transaction>;
  /** Type URL */
  typeUrl: Scalars['String']['output'];
};
// ...
export type Price = {
  __typename?: 'Price';
  amount: Scalars['String']['output'];
  /** Detected base denom, defaults to chain denom, null if equal to `denom` */
  baseDenom?: Maybe<Scalars['String']['output']>;
  currentUsd?: Maybe<PriceWithRate>;
  /** On chain denom specified by the chain, null if none specified */
  denom?: Maybe<Scalars['String']['output']>;
  /** Prefixed denom. For IBC denoms includes channels. */
  prefixedDenom?: Maybe<Scalars['String']['output']>;
  /** USD value at block time, if any, taken from coingecko */
  usd?: Maybe<PriceWithRate>;
};

export type PriceFilter = {
  max?: InputMaybe<Scalars['Decimal']['input']>;
  min?: InputMaybe<Scalars['Decimal']['input']>;
  useUsd?: Scalars['Boolean']['input'];
};

export type PriceWithRate = {
  __typename?: 'PriceWithRate';
  amount: Scalars['Decimal']['output'];
  coingeckoId?: Maybe<Scalars['String']['output']>;
  denom: Scalars['String']['output'];
  fromDenom: Scalars['String']['output'];
  rate: Scalars['Decimal']['output'];
  tracing?: Maybe<Scalars['String']['output']>;
};
// ...
export type Transaction = {
  __typename?: 'Transaction';
  blockHeight: Scalars['Int']['output'];
  code: Scalars['Int']['output'];
  createdAt: Scalars['DateTime']['output'];
  errorMessage?: Maybe<Scalars['String']['output']>;
  /** Careful: this is only returning the last 100 events */
  events?: Maybe<Array<Event>>;
  gasUsed: Scalars['Int']['output'];
  gasWanted: Scalars['Int']['output'];
  /** Transaction Hash */
  hash: Scalars['String']['output'];
  /** Internal UUID, might change between when the indexer is reran. */
  id: Scalars['ID']['output'];
  idx: Scalars['Int']['output'];
  memo?: Maybe<Scalars['String']['output']>;
  /** Careful: this is only returning the last 100 messages */
  messages?: Maybe<Array<Message>>;
};

export type TransactionConnection = {
  __typename?: 'TransactionConnection';
  edges: Array<TransactionEdge>;
  pageInfo: PageInfo;
};

export type TransactionEdge = {
  __typename?: 'TransactionEdge';
  cursor: Scalars['String']['output'];
  node: Transaction;
};

export enum TransactionSortBy {
  BlockHeightAsc = 'BLOCK_HEIGHT_ASC',
  BlockHeightDesc = 'BLOCK_HEIGHT_DESC'
}

export type Address = {
  __typename?: 'address';
  addr: Scalars['String']['output'];
  /** the associated name, if any */
  name?: Maybe<Name>;
};

export type Collection = {
  __typename?: 'collection';
  /** Is this collection manually blocked */
  blocked: Scalars['Boolean']['output'];
  /** Collection address */
  collectionAddr: Scalars['String']['output'];
  /**
   * When was this collection created at.
   * @deprecated Field created_at is deprecated, use minted_at instead
   */
  createdAt: Scalars['DateTime']['output'];
  /** The address that created this collection */
  createdBy: Address;
  /** Creator address */
  createdByAddr: Scalars['String']['output'];
  /** Description */
  description: Scalars['String']['output'];
  /**
   * The floor price is the lowest price for any currently for sale item in the collection.
   * Returned as STARS.
   */
  floorPrice?: Maybe<Scalars['Int']['output']>;
  /**
   * The floor price is the lowest price for any currently for sale item in the collection.
   * Returned as USD.
   */
  floorPriceUsd?: Maybe<Scalars['Decimal']['output']>;
  /** Internal UUID, might change between when the indexer is reran. */
  id: Scalars['ID']['output'];
  /** Main image */
  image: Scalars['String']['output'];
  imageObject?: Maybe<Image>;
  /** When was this collection minted on chain. UTC without timezone */
  mintedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Collection minter address */
  minterAddr?: Maybe<Scalars['String']['output']>;
  /** Name */
  name: Scalars['String']['output'];
  /**
   * How many tokens this collection has for the given owner when called through
   * OwnedCollections
   */
  ownerTokensCount?: Maybe<Scalars['Int']['output']>;
  /** Owners for this collection */
  owners: Collection_Owners;
  ownersCount?: Maybe<Scalars['Int']['output']>;
  /** How many tokens this collection has */
  tokensCount?: Maybe<Scalars['Int']['output']>;
  /** Website */
  website?: Maybe<Scalars['String']['output']>;
  /** Whitelist address */
  whitelistAddr?: Maybe<Scalars['String']['output']>;
};


export type CollectionOwnersArgs = {
  limit?: InputMaybe<Scalars['Int']['input']>;
  offset?: InputMaybe<Scalars['Int']['input']>;
};

export type Collection_Node = {
  __typename?: 'collection_node';
  collections: Array<Collection>;
  limit: Scalars['Int']['output'];
  offset: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};

export type Collection_Owner = {
  __typename?: 'collection_owner';
  count: Scalars['Int']['output'];
  owner: Address;
};

export type Collection_Owners = {
  __typename?: 'collection_owners';
  owners: Array<Collection_Owner>;
  totalCount: Scalars['Int']['output'];
};
// ...
export type Contract = {
  __typename?: 'contract';
  blockHeight: Scalars['Int']['output'];
  blocked: Scalars['Boolean']['output'];
  contractAddr: Scalars['String']['output'];
  contractCodeId?: Maybe<Scalars['String']['output']>;
  contractInfo: Scalars['JSONObject']['output'];
  contractLabel?: Maybe<Scalars['String']['output']>;
  contractType: Scalars['String']['output'];
  contractVersion?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['NaiveDateTime']['output'];
  createdMessageId?: Maybe<Scalars['UUID']['output']>;
  lastErrorAt?: Maybe<Scalars['NaiveDateTime']['output']>;
  updatedAt: Scalars['NaiveDateTime']['output'];
};
// ...
export type Image = {
  __typename?: 'image';
  contentLength?: Maybe<Scalars['Int']['output']>;
  contentType?: Maybe<Scalars['String']['output']>;
  height?: Maybe<Scalars['Int']['output']>;
  url: Scalars['String']['output'];
  width?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Name = {
  __typename?: 'name';
  askPrice?: Maybe<Price>;
  associatedAddr?: Maybe<Scalars['String']['output']>;
  contractAddr: Scalars['String']['output'];
  forSale: Scalars['Boolean']['output'];
  highestOffer?: Maybe<Scalars['Decimal']['output']>;
  highestOfferEvent?: Maybe<Event>;
  id: Scalars['ID']['output'];
  imageUrl?: Maybe<Scalars['String']['output']>;
  mintedAt?: Maybe<Scalars['DateTime']['output']>;
  name: Scalars['String']['output'];
  ownerAddr: Scalars['String']['output'];
  records?: Maybe<Array<Name_Record>>;
};
// ...
export type Token = {
  __typename?: 'token';
  animation?: Maybe<Image>;
  animationUrl?: Maybe<Scalars['String']['output']>;
  askId?: Maybe<Scalars['String']['output']>;
  collectionAddr: Scalars['String']['output'];
  createdAt: Scalars['DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  escrowContractAddr?: Maybe<Scalars['String']['output']>;
  escrowContractType?: Maybe<Scalars['String']['output']>;
  forSale: Scalars['Boolean']['output'];
  /** The highest non-expired valid bid for this token */
  highestBid?: Maybe<Scalars['Decimal']['output']>;
  /** The highest non-expired valid bid for this token */
  highestBidEvent?: Maybe<Event>;
  /** The highest non-expired valid collection bid for the collection attached to this token */
  highestCollectionBid?: Maybe<Scalars['Decimal']['output']>;
  /** The highest non-expired valid collection bid for the collection attached to this token */
  highestCollectionBidEvent?: Maybe<Event>;
  highestLiveAuctionBidEvent?: Maybe<Event>;
  id: Scalars['ID']['output'];
  image?: Maybe<Image>;
  imageUrl?: Maybe<Scalars['String']['output']>;
  isEscrowed: Scalars['Boolean']['output'];
  /** The latest sale */
  lastSale?: Maybe<Event>;
  listedAt?: Maybe<Scalars['DateTime']['output']>;
  liveAuctionEndTime?: Maybe<Scalars['DateTime']['output']>;
  liveAuctionEvent?: Maybe<Event>;
  liveAuctionHighestBid?: Maybe<Price>;
  liveAuctionStartTime?: Maybe<Scalars['DateTime']['output']>;
  mintedAt?: Maybe<Scalars['DateTime']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  /** The address that owns this token */
  owner?: Maybe<Address>;
  ownerAddr?: Maybe<Scalars['String']['output']>;
  price?: Maybe<Price>;
  priceExpiresAt?: Maybe<Scalars['DateTime']['output']>;
  rarityOrder?: Maybe<Scalars['Int']['output']>;
  rarityScore?: Maybe<Scalars['Decimal']['output']>;
  saleType?: Maybe<Scalars['String']['output']>;
  sellerAddr?: Maybe<Scalars['String']['output']>;
  tokenId: Scalars['String']['output'];
  traits?: Maybe<Array<Token_Trait>>;
};

export type Token_Node = {
  __typename?: 'token_node';
  limit: Scalars['Int']['output'];
  offset: Scalars['Int']['output'];
  tokens: Array<Token>;
  total: Scalars['Int']['output'];
};

export type Token_Trait = {
  __typename?: 'token_trait';
  name: Scalars['String']['output'];
  rarity?: Maybe<Scalars['Float']['output']>;
  rarityPercent?: Maybe<Scalars['Float']['output']>;
  rarityScore?: Maybe<Scalars['Float']['output']>;
  value: Scalars['String']['output'];
};